apps/proofreader/requirements.txt
flask
anthropic
pdfplumber
python-pptx
python-docx
